+++ src/test/java/webserver/RequestParserTest.java
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.io.BufferedReader;
-import java.io.IOException;
 import java.io.StringReader;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -52,53 +51,4 @@ public class RequestParserTest {
         assertThat(httpRequest.getProtocol()).isEqualTo("HTTP/1.1");
     }
 
-    private static class RequestParser {
-        private final BufferedReader bufferedReader;
-
-        public RequestParser(BufferedReader bufferedReader) {
-            this.bufferedReader = bufferedReader;
-        }
-
-        public HttpRequest parse() {
-            return parseRequestLine();
-        }
-
-        private HttpRequest parseRequestLine() {
-            String requestLine = nextLine();
-            String[] requestLineSplit = requestLine.split(" ");
-            return new HttpRequest(requestLineSplit[0], requestLineSplit[1], requestLineSplit[2]);
-        }
-
-        private String nextLine() {
-            try {
-                return bufferedReader.readLine();
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
-        }
-    }
-
-    private static class HttpRequest {
-        private final String method;
-        private final String requestURI;
-        private final String protocol;
-
-        public HttpRequest(String method, String requestURI, String protocol) {
-            this.method = method;
-            this.requestURI = requestURI;
-            this.protocol = protocol;
-        }
-
-        public String getMethod() {
-            return method;
-        }
-
-        public String getRequestURI() {
-            return requestURI;
-        }
-
-        public String getProtocol() {
-            return protocol;
-        }
-    }
 }
